+++ bootique-jdbc-test/src/test/java/io/bootique/jdbc/test/TableIT.java
@@ -1,11 +1,20 @@
 package io.bootique.jdbc.test;
 
+import io.bootique.resource.ResourceFactory;
 import io.bootique.test.BQTestRuntime;
 import io.bootique.test.junit.BQTestFactory;
+import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+
+import java.sql.Date;
+import java.sql.Timestamp;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.util.Comparator;
+import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 
@@ -14,9 +23,10 @@ public class TableIT {
     @ClassRule
     public static BQTestFactory TEST_FACTORY = new BQTestFactory();
     private static Table T1;
+    private static Table T2;
 
     @Rule
-    public TestDataManager dataManager = new TestDataManager(true, T1);
+    public TestDataManager dataManager = new TestDataManager(true, T1, T2);
 
     @BeforeClass
     public static void setupDB() {
@@ -28,8 +38,10 @@ public class TableIT {
         DatabaseChannel channel = DatabaseChannel.get(testRuntime);
 
         channel.update("CREATE TABLE \"t1\" (\"c1\" INT, \"c2\" VARCHAR(10), \"c3\" VARCHAR(10))");
+        channel.update("CREATE TABLE \"t2\" (\"c1\" INT, \"c2\" INT, \"c3\" DATE, \"c4\" TIMESTAMP)");
 
         T1 = channel.newTable("t1").columnNames("c1", "c2", "c3").build();
+        T2 = channel.newTable("t2").columnNames("c1", "c2", "c3", "c4").build();
     }
 
     @Test
@@ -51,5 +63,58 @@ public class TableIT {
         assertEquals(0, T1.getRowCount());
         T1.insertColumns("c2", "c1").values("v1", 1).values("v2", 2).exec();
         assertEquals(2, T1.getRowCount());
+    }
+
+    @Test
+    public void testInsertFromCsv_Empty() {
+        assertEquals(0, T1.getRowCount());
+        T1.insertFromCsv(new ResourceFactory("classpath:io/bootique/jdbc/test/empty.csv"));
+        assertEquals(0, T1.getRowCount());
+    }
+
+    @Test
+    public void testInsertFromCsv() {
+        assertEquals(0, T1.getRowCount());
+        T1.insertFromCsv(new ResourceFactory("classpath:io/bootique/jdbc/test/t1.csv"));
+
+        List<Object[]> data = T1.select();
+        assertEquals(2, data.size());
+
+        // sort in memory, as there's no guarantee that DB will return data in insertion order
+        data.sort(Comparator.comparing(r -> (Integer) r[0]));
+
+        Object[] row1 = data.get(0);
+        assertEquals(1, row1[0]);
+        Assert.assertNull(row1[1]);
+        assertEquals("abcd", row1[2]);
+
+        Object[] row2 = data.get(1);
+        assertEquals(2, row2[0]);
+        assertEquals("tt", row2[1]);
+        assertEquals("xyz", row2[2]);
+    }
+
+    @Test
+    public void testInsertFromCsv_Dates() {
+        assertEquals(0, T2.getRowCount());
+        T2.insertFromCsv(new ResourceFactory("classpath:io/bootique/jdbc/test/t2.csv"));
+
+        List<Object[]> data = T2.select();
+        assertEquals(2, data.size());
+
+        // sort in memory, as there's no guarantee that DB will return data in insertion order
+        data.sort(Comparator.comparing(r -> (Integer) r[0]));
+
+        Object[] row1 = data.get(0);
+        assertEquals(1, row1[0]);
+        Assert.assertNull(row1[1]);
+        assertEquals(Date.valueOf(LocalDate.of(2016, 1, 9)), row1[2]);
+        assertEquals(Timestamp.valueOf(LocalDateTime.of(2016, 1, 10, 10, 0, 0)), row1[3]);
+
+        Object[] row2 = data.get(1);
+        Assert.assertNull(row2[1]);
+        assertEquals(Date.valueOf(LocalDate.of(2017, 1, 9)), row2[2]);
+        assertEquals(Timestamp.valueOf(LocalDateTime.of(2017, 1, 10, 13, 0, 1)), row2[3]);
+
     }
 }